Add files via upload
Object Array of letters with a Jitter effect. 
Exploring the idea of the user creating the graphics, and then having the graphics come "alive".
Moving forward would like to possibly add a GUI to better control the jitter and giving the user more customization. Possibly adding a complete alphabet and allowing the user to type their own words to have come to life.